update MSSQL, MSSQL MS Data multi-result set tests: - label lib feature (MultiResultsSet) & issues from github - tests moved to BaseQueryTests class
