+++ PetaPoco.Tests.Integration/Databases/BaseQueryTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -8,8 +8,6 @@ using PetaPoco.Tests.Integration.Models;
 using PetaPoco.Utilities;
 using Shouldly;
 using Xunit;
-
-// TODO: Finish labeling DB Features for tests with provider-specific features.
 
 namespace PetaPoco.Tests.Integration.Databases
 {
@@ -1430,6 +1428,148 @@ namespace PetaPoco.Tests.Integration.Databases
 
             var results = await DB.SkipTakeAsync<string>(2, 1, sql);
             results.Count.ShouldBe(1);
+        }
+
+        [Fact]
+        [Trait("LibFeature", "MultiResultsSet")]
+        public virtual void Query_MultiResultsSet_SingleResultsSetSinglePoco__ShouldReturnValidPocoCollection()
+        {
+            AddPeople(1, 0);
+            var pd = PocoData.ForType(typeof(Person), DB.DefaultMapper);
+            var sql = "SET CONCAT_NULL_YIELDS_NULL ON;" +
+                      $"SELECT * FROM [{pd.TableInfo.TableName}] WHERE [{pd.Columns.Values.First(c => c.PropertyInfo.Name == "Name").ColumnName}] LIKE @0 + '%'";
+
+            List<Person> result;
+            using (var multi = DB.QueryMultiple(sql, "Peta"))
+            {
+                result = multi.Read<Person>().ToList();
+            }
+
+            result.Count.ShouldBe(1);
+            var person = result.First();
+
+            person.Id.ShouldNotBe(Guid.Empty);
+            person.Name.ShouldStartWith("Peta");
+            person.Age.ShouldBe(18);
+        }
+
+        [Fact]
+        [Trait("LibFeature", "MultiResultsSet")]
+        public virtual void Query_MultiResultsSet_SingleResultsSetMultiPoco__ShouldReturnValidPocoCollection()
+        {
+            AddOrders(1);
+
+            var pd = PocoData.ForType(typeof(Person), DB.DefaultMapper);
+            var od = PocoData.ForType(typeof(Order), DB.DefaultMapper);
+            var sql = "SET CONCAT_NULL_YIELDS_NULL ON;" +
+                      $"SELECT TOP 1 * FROM [{od.TableInfo.TableName}] o INNER JOIN [{pd.TableInfo.TableName}] p ON p.[{pd.Columns.Values.First(p => p.PropertyInfo.Name == "Id").ColumnName}] = o.[{od.Columns.Values.First(p => p.PropertyInfo.Name == "PersonId").ColumnName}] WHERE [{pd.Columns.Values.First(c => c.PropertyInfo.Name == "Name").ColumnName}] = @0 ORDER BY 1 DESC";
+
+            List<Order> result;
+            using (var multi = DB.QueryMultiple(sql, "Peta0"))
+            {
+                result = multi.Read<Order, Person, Order>((o, p) =>
+                {
+                    o.Person = p;
+                    return o;
+                }).ToList();
+            }
+
+            result.Count.ShouldBe(1);
+            var order = result.First();
+
+            order.PoNumber.ShouldStartWith("PO");
+            order.Status.ShouldBeOneOf(Enum.GetValues(typeof(OrderStatus)).Cast<OrderStatus>().ToArray());
+            order.PersonId.ShouldNotBe(Guid.Empty);
+            order.CreatedOn.ShouldBeLessThanOrEqualTo(new DateTime(1990, 1, 1, 0, 0, 0, DateTimeKind.Utc));
+            order.CreatedBy.ShouldStartWith("Harry");
+
+            order.Person.ShouldNotBeNull();
+            order.Person.Id.ShouldNotBe(Guid.Empty);
+            order.Person.Name.ShouldStartWith("Peta");
+            order.Person.Age.ShouldBe(18);
+        }
+
+        [Fact]
+        [Trait("LibFeature", "MultiResultsSet")]
+        public virtual void Query_MultiResultsSet_MultiResultSetSinglePoco__ShouldReturnValidPocoCollection()
+        {
+            AddOrders(1);
+
+            var pd = PocoData.ForType(typeof(Person), DB.DefaultMapper);
+            var od = PocoData.ForType(typeof(Order), DB.DefaultMapper);
+            var sql =
+                $"SET CONCAT_NULL_YIELDS_NULL ON;SELECT * FROM [{od.TableInfo.TableName}] o WHERE [{od.Columns.Values.First(c => c.PropertyInfo.Name == "Id").ColumnName}] = @0;SELECT * FROM [{pd.TableInfo.TableName}] p WHERE [{pd.Columns.Values.First(c => c.PropertyInfo.Name == "Name").ColumnName}] = @1;";
+
+            Order order;
+            using (var multi = DB.QueryMultiple(sql, "1", "Peta0"))
+            {
+                order = multi.Read<Order>().First();
+                order.Person = multi.Read<Person>().First();
+            }
+
+            order.PoNumber.ShouldStartWith("PO");
+            order.Status.ShouldBeOneOf(Enum.GetValues(typeof(OrderStatus)).Cast<OrderStatus>().ToArray());
+            order.PersonId.ShouldNotBe(Guid.Empty);
+            order.CreatedOn.ShouldBeLessThanOrEqualTo(new DateTime(1990, 1, 1, 0, 0, 0, DateTimeKind.Utc));
+            order.CreatedBy.ShouldStartWith("Harry");
+
+            order.Person.ShouldNotBeNull();
+            order.Person.Id.ShouldNotBe(Guid.Empty);
+            order.Person.Name.ShouldStartWith("Peta");
+            order.Person.Age.ShouldBe(18);
+        }
+
+        [Fact]
+        [Trait("LibFeature", "MultiResultsSet")]
+        public virtual void Query_MultiResultsSet_MultiResultSetMultiPoco__ShouldReturnValidPocoCollection()
+        {
+            AddOrders(12);
+
+            var pd = PocoData.ForType(typeof(Person), DB.DefaultMapper);
+            var od = PocoData.ForType(typeof(Order), DB.DefaultMapper);
+            var old = PocoData.ForType(typeof(OrderLine), DB.DefaultMapper);
+            var sql = "SET CONCAT_NULL_YIELDS_NULL ON;" +
+                      $"SELECT * FROM [{od.TableInfo.TableName}] o INNER JOIN [{pd.TableInfo.TableName}] p ON p.[{pd.Columns.Values.First(p => p.PropertyInfo.Name == "Id").ColumnName}] = o.[{od.Columns.Values.First(p => p.PropertyInfo.Name == "PersonId").ColumnName}] ORDER BY o.[{od.Columns.Values.First(p => p.PropertyInfo.Name == "Id").ColumnName}] ASC;SELECT * FROM [{old.TableInfo.TableName}] ol ORDER BY ol.[{old.Columns.Values.First(c => c.PropertyInfo.Name == "OrderId").ColumnName}] ASC";
+
+            List<Order> results;
+            using (var multi = DB.QueryMultiple(sql))
+            {
+                results = multi.Read<Order, Person, Order>((o, p) =>
+                {
+                    o.Person = p;
+                    return o;
+                }).ToList();
+                var orderLines = multi.Read<OrderLine>().ToList();
+
+                foreach (var order in results)
+                {
+                    order.OrderLines = orderLines.Where(ol => ol.OrderId == order.Id).ToList();
+                }
+            }
+
+            results.Count.ShouldBe(12);
+
+            results.ForEach(o =>
+            {
+                o.PoNumber.ShouldStartWith("PO");
+                o.Status.ShouldBeOneOf(Enum.GetValues(typeof(OrderStatus)).Cast<OrderStatus>().ToArray());
+                o.PersonId.ShouldNotBe(Guid.Empty);
+                o.CreatedOn.ShouldBeLessThanOrEqualTo(new DateTime(1990, 1, 1, 0, 0, 0, DateTimeKind.Utc));
+                o.CreatedBy.ShouldStartWith("Harry");
+
+                o.Person.ShouldNotBeNull();
+                o.Person.Id.ShouldNotBe(Guid.Empty);
+                o.Person.Name.ShouldStartWith("Peta");
+                o.Person.Age.ShouldBeGreaterThanOrEqualTo(18);
+
+                o.OrderLines.Count.ShouldBe(2);
+                var firstOrderLine = o.OrderLines.First();
+                firstOrderLine.Quantity.ToString().ShouldBe("1");
+                firstOrderLine.SellPrice.ShouldBe(9.99m);
+                var secondOrderLine = o.OrderLines.Skip(1).First();
+                secondOrderLine.Quantity.ToString().ShouldBe("2");
+                secondOrderLine.SellPrice.ShouldBe(19.98m);
+            });
         }
     }
 }
